refactor(av1): fix Clippy warnings in overlay extraction
- Remove redundant else-if block in mv_extractor (both branches identical)
- Remove unnecessary usize casting in partition.rs
- Derive Default for FrameTypeInfo instead of manual impl

This simplifies the code and follows Rust best practices.

+++ crates/bitvue-av1/src/overlay_extraction/partition.rs
@@ -407,7 +407,7 @@ fn get_intra_mode_for_position(col: u32, row: u32) -> PredictionMode {
         PredictionMode::PaethPred,
     ];
 
-    let idx = (((col as usize) + (row as usize) * 3) % INTRA_MODES.len()) as usize;
+    let idx = ((col as usize) + (row as usize) * 3) % INTRA_MODES.len();
     INTRA_MODES[idx]
 }
 
@@ -420,7 +420,7 @@ fn get_inter_mode_for_position(col: u32, row: u32) -> PredictionMode {
         PredictionMode::GlobalMv,
     ];
 
-    let idx = (((col as usize) + (row as usize)) % INTER_MODES.len()) as usize;
+    let idx = ((col as usize) + (row as usize)) % INTER_MODES.len();
     INTER_MODES[idx]
 }
